+++ src/index.js
@@ -1,4 +1,5 @@
 import BaseRepository from './base-repository';
+import createController from './create-controller';
 import * as constants from './constants';
 
-export {BaseRepository, constants};
+export {BaseRepository, constants, createController};